30-DEC

+++ catalog/rest/rest_views.py
@@ -8,11 +8,11 @@ from catalog.models import Customer
 class CustomerSerializer(serializers.ModelSerializer):
     class Meta:
         model = Customer
-        fields = ('id', 'name', 'email', 'mobile','profile')
+        fields = ('id', 'name', 'email', 'mobile', 'remarks')
 
 
 @api_view(['GET', 'POST'])
-def process_Customers(request):
+def process_customers(request):
     if request.method == "GET":
         customers = Customer.objects.all()
         serializer = CustomerSerializer(customers, many=True)
@@ -26,8 +26,8 @@ def process_Customers(request):
             return Response(customer.errors, status=400)  # bad request
 
 
-@api_view(['GET', 'DELETE'])
-def process_Customer(request, id):
+@api_view(['GET', 'DELETE', 'PUT'])
+def process_one_customer(request, id):
     try:
         customer = Customer.objects.get(id=id)
     except:
@@ -36,6 +36,13 @@ def process_Customer(request, id):
     if request.method == "GET":
         serializer = CustomerSerializer(customer)
         return Response(serializer.data)
-    else:  # DELETE
+    elif request.method == "DELETE":  # DELETE
         customer.delete()
         return Response(status=204)  # No data
+    else:
+        email = request.POST['email']
+        customer.email = email
+        customer.save()  # Update table in DB
+        serializer = CustomerSerializer(customer)
+        return Response(serializer.data)
+
